refactor: MVC 패턴에 따라서 패키지 정리
