+++ src/main/java/lotto/model/domain/LottoMoney.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model.domain;
 
 public class LottoMoney {
