+++ src/main/java/lotto/model/domain/LottoNumber.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model.domain;
 
 import java.util.Objects;
