+++ src/main/java/lotto/model/domain/RandomNumbersGenerator.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model.domain;
 
 import java.util.ArrayList;
 import java.util.Collections;
