+++ src/main/java/lotto/model/domain/Rank.java
@@ -1,6 +1,5 @@
-package lotto;
+package lotto.model.domain;
 
-import java.lang.reflect.Array;
 import java.util.Arrays;
 
 public enum Rank {
