+++ src/test/java/lotto/model/domain/LottoMoneyTest.java
@@ -1,6 +1,7 @@
-package lotto;
+package lotto.model.domain;
 
 
+import lotto.model.domain.LottoMoney;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
