+++ src/test/java/lotto/model/domain/LottosTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model.domain;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
