+++ src/test/java/lotto/model/domain/RandomNumbersGeneratorTest.java
@@ -1,8 +1,8 @@
-package lotto;
+package lotto.model.domain;
 
+import lotto.model.domain.RandomNumbersGenerator;
 import org.junit.jupiter.api.Test;
 
-import java.util.HashSet;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
