+++ src/test/java/lotto/model/domain/RankTest.java
@@ -1,9 +1,9 @@
-package lotto;
+package lotto.model.domain;
 
 
+import lotto.model.domain.Rank;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.MethodSource;
 
 import java.util.stream.Stream;
